+++ release-notes.md
@@ -1,5 +1,5 @@
 ## @thu-info/app
-- 支持展示任意时刻的日程
+- 支持上传自定义日程
 
 ## @thu-info/lib
-- TBA
+- feat: implement `saveCustomSchedule`